fix: suggestions from comments
