+++ test/ae_mdw/aexn_tokens_test.exs
@@ -234,61 +234,5 @@ defmodule AeMdw.AexnTokensTest do
       assert {:ok, {nil, [^m1, ^m2], nil}} =
                AexnTokens.fetch_contracts(state, pagination, :aex141, %{}, :creation, nil)
     end
-
-    test "return AEX-9 contract, filtering the invalid ones" do
-      contract_pk1 = <<2::256>>
-      contract_pk2 = <<1::256>>
-      txi1 = 123_456_788
-      txi2 = 123_456_789
-
-      m1 =
-        Model.aexn_contract(
-          index: {:aex9, contract_pk1},
-          txi_idx: {txi1, -1},
-          meta_info: {"TokenB1", "TKB1", 18},
-          extensions: ["extB1"]
-        )
-
-      m2 =
-        Model.aexn_contract(
-          index: {:aex9, contract_pk2},
-          txi_idx: {txi2, -1},
-          meta_info: {"TokenA1", "TKA1", 18},
-          extensions: ["extA1"]
-        )
-
-      state =
-        empty_state()
-        |> State.put(Model.AexnContract, m1)
-        |> State.put(Model.AexnContract, m2)
-        |> State.put(
-          Model.AexnContractCreation,
-          Model.aexn_contract_creation(index: {:aex9, {txi1, -1}}, contract_pk: contract_pk1)
-        )
-        |> State.put(
-          Model.AexnContractCreation,
-          Model.aexn_contract_creation(index: {:aex9, {txi2, -1}}, contract_pk: contract_pk2)
-        )
-        |> State.put(
-          Model.AexnContractName,
-          Model.aexn_contract_name(index: {:aex9, "TokenB1", contract_pk1})
-        )
-        |> State.put(
-          Model.AexnContractName,
-          Model.aexn_contract_name(index: {:aex9, "TokenA1", contract_pk2})
-        )
-        |> State.put(
-          Model.AexnInvalidContract,
-          Model.aexn_invalid_contract(
-            index: {:aex9, contract_pk1},
-            reason: Aex9.invalid_number_of_holders()
-          )
-        )
-
-      pagination = {:forward, false, 10, false}
-
-      assert {:ok, {nil, [^m2], nil}} =
-               AexnTokens.fetch_contracts(state, pagination, :aex9, %{}, :name, nil)
-    end
   end
 end
